refact: remove task from a daily plan
