+++ apps/web/app/api/daily-plan/plan/[planId]/route.ts
@@ -1,6 +1,6 @@
 import { IDailyPlanTasksUpdate, INextParams } from '@app/interfaces';
 import { authenticatedGuard } from '@app/services/server/guards/authenticated-guard-app';
-import { addTaskToDailyPlanRequest } from '@app/services/server/requests';
+import { addTaskToDailyPlanRequest, removeTaskFromPlanRequest } from '@app/services/server/requests';
 import { NextResponse } from 'next/server';
 
 export async function POST(req: Request, { params }: INextParams) {
@@ -21,6 +21,29 @@ export async function POST(req: Request, { params }: INextParams) {
 		data: body,
 		planId,
 		tenantId
+	});
+
+	return $res(response.data);
+}
+
+export async function PUT(req: Request, { params }: INextParams) {
+	const res = new NextResponse();
+
+	const { planId } = params;
+	if (!planId) {
+		return;
+	}
+
+	const { $res, user, tenantId, access_token } = await authenticatedGuard(req, res);
+	if (!user) return $res('Unauthorized');
+
+	const body = (await req.json()) as unknown as IDailyPlanTasksUpdate;
+
+	const response = await removeTaskFromPlanRequest({
+		data: body,
+		planId,
+		tenantId,
+		bearer_token: access_token
 	});
 
 	return $res(response.data);
